backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
notion-client
python-multipart
reportlab
emoji
